Create temporary images and graphics as needed
Create the various types of intermediate AWT images and associated
graphics as needed in each test function, allowing for simpler variable
names within their more limited scope. Rename the loaded PNG image to
"pngImage" so that "awtImage" refers to the intermediate AWT image,
"jfxImage" refers to a JavaFX image that uses a PixelWriter, and
"nioImage" refers to a JavaFX image that uses a PixelBuffer.