Do not clear the AWT background to a solid color
Clearing the background to a solid color before drawing the PNG image
results in an AWT image with no transparency, correctly blended into the
background. When such an image is written to the JavaFX image, it has
the same correct appearance whether or not the specified pixel format is
premultiplied.

Instead, leave the background transparent so that the final AWT image
also has transparency. Such an image must be blended when written to the
JavaFX image, and its appearance will differ depending on whether the
specified pixel format is premultiplied.